refactor(transaction-service): cached-registry-client - phase 2: extract keyv store factory
